IOS-3925 Move file
